Custom Spreadsheet Software
This program simulates a collaborative real-time spreadsheet editor
similar to Google Spreadsheet. The program operates on an input and
output text file. The input file contains a formatted list of the
current documents, corresponding users and also operations to perform on
the database. The server than writes the resulting to an output.txt file
after every operation is applied to the database. The allowed operations
include set, clear, add, subtract, multiply, divide, undo and redo.